split session tests
